Update create-automation-tasks-azure-resources.md
I am working as an Azure Cloud Engineer in Microsoft. Recently I received a customer support request where the customer was complaining that the auto start did not work as expected. But probably he was unclear about scheduling auto start task for his VM due to lack of documentation.
He wanted to scheduled an auto start task for his VM once every day for a month. But he observed that the VM auto started only for 1 day and did not start the next day onwards. While troubleshooting I observed that he had mentioned "1" in the Interval tab and "Month" in the Frequency tab which suggests the auto start to work once in a month and hence it did not start everyday.

Configuring automation for Start/stop VM is a very common requirement for the customers. Since currently Tasks feature is in Preview, many customers can get confused, similar to my customer and end up opening tickets. As I do not see any documents available specific to scheduling auto start/stop task for the VM, I wanted to publish a document from scratch for creating Automation tasks for auto start/stop for a VM, but I am not sure about the process to do that. Hence I have just updated the existing document so that customers can easily identify how the task scheduling works.

Please review it and provide your feedback.