Minor fixes on issued group/roles claims
When "SecurityGroup" is set we should inform that both security groups and Azure AD Roles will be issued as it's described in https://learn.microsoft.com/en-us/azure/active-directory/develop/reference-app-manifest#groupmembershipclaims-attribute.  

When "SecurityGroup" is set in conjunction with Group filtering, since both security groups and Azure AD Roles are issued, there is a common misinterpretation that Group filtering will also be applied to Azure AD Roles, when it is not. Updating documentation accordingly.