Update how-provisioning-works.md
Removed "Use" from the hyperlink "Use delta query to track changes in Microsoft Graph data" to suit the flow of the sentence in the following block:-

When Azure AD is the source system, the provisioning service uses the Use delta query to track changes in Microsoft Graph data to monitor users and groups. The provisioning service runs an initial cycle against the source system and target system, followed by periodic incremental cycles.