Update trusted-launch.md
Changes in diction include elimination of the word "limitation" - it is confusing to the reader instead changing the VM Size and OS Size to be headers while adding an "additional information" for the region, pricing, and coming soon features to branch under. A change in the view of the VM Sizes Supported, to a table to better visualize supported SKUs and adding a link to gen 2 page to see comparisons. OS View Change to table.