Tweak name to avoid confusion with ToU
I ASSUME the section being updated in this PR was referring to https://learn.microsoft.com/en-us/graph/api/resources/privacyprofile?view=graph-rest-1.0 , hence updating it to avoid confusion with terms of use that is already mentioned in the Conditional Access section. That said, the "Organizational relationships" section may be a little confusing in general?