Update translator-how-to-install-container.md
Per Jonathan:
"The hardware requirements apply to both standard and disconnected containers so making this general as it has come up with customers. We don't really use the term "connected" in our public facing container docs either, we simply refer to containers and differentiate containers that can be run without an internet connection as disconnected."

We can also bump up the version or use "latest" for future proofing.