(AB#85432) Rename machine configuration articles
Prior to this change, nearly all of the machine configuration articles were
named with the `machine-configuration-` prefix. They are all located in the
`machine-configuration` folder, which made their slug url repeat the phrase
`machine-configuration` twice in a row.

This change renames the machine configuration articles to make them easier to
understand when just reading the names in the repository and to fix the repeated
phrase problem in the article slugs.

It updates all references inside the folder and adds redirects.