Update traffic-analytics-schema.md
Removed reference in schema table to field 'PublicIP_s'. This was identified as part of case 2304110060000554. On further investigation, it was found that this field was not supported and is likely a typo. To confirm, code history was checked by Niti Gupta and mentioned that no historical evidence of this field is present