Update connector-postgresql.md
The path mentioned for the rootcert for the ODBC to work C:\users\DIAHostService\AppData\Roaming\postgresql\root.crt is not work .  DIAHostservice is a service account and it exist in the C:\Windows\ServiceProfiles. So the correct path will be C:\Windows\ServiceProfiles\DIAHostService\AppData\Roaming\postgresql\root.crt.