Add information about customer-managed keys (CMK)
As it was confirmed by Chris H. earlier today in a group chat, CMK does encrypt the fine-tuned models. That's why as discussed with John S., adding PR for the reference to CMK as customers are not sure whether it's applicable only to training and validation data, or fine-tuned models as well. Thank you.