Add information about exporting B2C Consumer users and how UPN info is exhibited
Suggestion to add the information on how to export a csv file with the B2C Consumer users and also the important information about how UPN information is brought into that file. UPN information come with the objectID@B2CDomain format and might cause some confusion to the customers as this is expected.