Update soft-delete-blob-overview.md
Updated the explanation on performing an Undelete operation on a soft-deleted directory. The current explanation caused much confusion because it makes it sound like if you rename the current version of the directory, you will "disconnect" the blobs from the directory and be unable to restore the soft-deleted directory and its contents. When in reality, you are required to rename the current version of the directory if you want to retain it while also Undeleting the original version of the directory.