Clarification for Subnet Delegation
We see several failures that customers experience when they enable subnet delegation for subnets where APIM will be deployed. This is a bit counter intuitive - as when customers create subnet from Portal, they have a tendency to enable delegation for Microsoft.ApiManagement. This causes a deployment failure that generates a support case. This clarity should help reduce that confusion.