Update manage-data-sources.md
"Your collection has been moved" should be "Your data source has been moved". The data source is the object that is moved to the selected collection. The collection itself is not moved!

Confusion arises because the window in the previous step is called "Move collection". Apparently one and the same window is used for moving different asset types, so a generic name has been attached to it. It would have been clearer if the window had been called "Move data source" because you don't move the collection but the data source...