Update compute-infrastructure.md
As Qatar Central is a recently established Azure region, only the stv2 platform is supported for API Management services deployed in this region.

I am from the APIM Support team. I had verified this piece of information with the Product Team to cross check as well. 

Updating this document for customer's reference.