Update connection-latency.md
Added an extra detail about AVD Gateway as follows - In some instances, the gateway chosen for a user connection may not be the closest to the user's location since the selection is determined based on the latency and it could be possible that a nearby region offers lower latency. Additionally, this situation may arise due to the client's network configuration, such as the use of a VPN Gateway, or when Azure Front Door experiences capacity issues with gateway servers in specific regions.