AKS still doesn't support AutoML jobs
As confirmed by BK in the internal AVA, AKS currently doesn't support AutoML jobs.

This part of documentation is not correct and conflicts with this correct statement (https://learn.microsoft.com/en-gb/azure/machine-learning/how-to-configure-auto-train?view=azureml-api-2#compute-to-run-experiment), which indicates that only compute instances and compute clusters can be used in SDK v2 and CLI v2 for AutoML.

In the same way, when I open AutoML in portal UI, it doesn't show Kubernetes as supported "compute type". Thanks.