Update activity-log.md
It appears to be a common user pain point that fields in AzureActivity do not always have the same capitalization for equivalent values contained within them. This can be worked around with Kusto queries designed specifically to account for this, but there is no currently listed guidance for this related to the Activity Log in Log Analytics specifically.