Update how-to-connect-fed-group-claims.md
Regex transform is a feature under group transformation. So it works only when the user is a member of less than 150 groups for SAML. A customer complained about this document as it is not clear enough. The customer tried to use regex transform to avoid the number of groups limit but of course it didn't work. The customer complained that this document is not clear with that. So I'm proposing this change.