Update partner-xid.md
In the following document, there is no mention of being able to use the 'METADATA' tag within the 'OAuth2 Technical Profile' metadata. 
https://learn.microsoft.com/en-us/azure/active-directory-b2c/oauth2-technical-profile

I've also received feedback from customers who encountered a '403' error that prevented operation. They reported that it worked correctly when using the 'authorization_endpoint' parameter. If you are to use the 'metadata', I believe that the 'OIDC Technical Profile' would be the correct one, hence I've submitted a Pull Request.