Update hdinsight-restrict-public-connectivity.md
In the testing that I did for a MAG implementation, the ARM template configured with the outbound value did not deploy with an internal error UNLESS the inbound NSG rule and UDR routing back to the internet was setup. 

With this in place the implementation worked as planned and was a great security improvement with integrated vnet configuration. If you don't like the phrasing please disregard. Wanted to provide the information for those in the same situation.