Update users-default-permissions.md
Enterprise Application Owners can manage provisioning settings for the Service Principals they own, but this is not reflected in the actions listed for Owned enterprise applications:

Enterprise application owner permissions

When a user adds a new enterprise application, they're automatically added as an owner. As an owner, they can manage the tenant-specific configuration of the application, such as the SSO configuration, PROVISIONING, and user assignments.