Update faq.yml with region availability of the service
We have added a link to the ACOM product by region page and also added the same regions list here as we are getting customer inquiries about the availability of the product in Azure regions. We will continue to update this list as we deploy the product to more regions.