Update reference-rest-api-data-plane.md
Fix the incorrect description of the contents of the JWT's audience. It is in
fact the exact same URL as the request URL - the query and the trailing slash
is kept, as can be seen in the official SDKs.

Further, the example is wrong. The broadcast URL ends with /:send and also
requires the api-version query parameter to be set.