Update defender-for-cloud-planning-and-operations-guide.md
Shortened or tightened various sections as well as added content for clarity.
Added links and anchors.
Reorganized some materials, including moving some content into tables.
Added additional alt text.

Outstanding questions: 
For the example AZURE RBAC roles, I left the display of Owner/Contributor but it would be good to confirm that these are either/or and not adding another role. I would probably ask if this could be written with commas instead. 

For the "Sam" example confirmation for the workspace permissions.
Confirmation that only subscription and resource group Owners and Contributors can apply security recommendations for a resource. 

Is my understanding of lines 100-107 is correct in that these are areas that contribute to creating a security policy?

Does Azure monitor agent also get installed automatically with cloud defender?

Which currently gets installed AMA or Log Analytics agent? 

Are the descriptions for log analytics still relevant to AMA?