Update how-to-configure-shir-for-log-analytics-collection.md
There were long hyphens in the name "Connectors - Integration Runtime".  The long hyphen when copied by a customer to create the integration connector results in Log Analytics MMA agent to fail processing the MMA configuration file which is only present for ~10 milliseconds making the problem very difficult determine.  I also noticed the same long hyphen in the Kusto query so removed it from there also as that will not work as expected either.