(MAINT) Fix wording for deleted policy requirements
Prior to this change, the section describing required actions when
an Azure Policy assignment is deleted used the same leading text
for each paragraph, even though they describe different required
actions.

This is because a critical phrase was erroneously removed from the
second paragraph, distinguishing what users need to do when an
Azure Policy assignment is deleted _from an initiative_.

This change adds the phrase back into the document.