Update functions-reference-node.md
The Linux tab in the Node version upgrade section says "When running on **Windows**, the Node.js version is set by the linuxfxversion site setting. This setting can be updated using the Azure CLI."

Correcting this to:
"When running on Linux, the Node.js version is set by the linuxfxversion site setting. This setting can be updated using the Azure CLI."