Replace 'System' by 'system' to use System-Managed Identity to access key vault secret
Replace 'System' by 'system' to use System-Managed Identity to access key vault secret. When creating in the portal and exporting the template, it uses 'system'. When creating with 'System' it is not recognized in the portal when the secret is edited. For Dapr component, it is 'system' and not 'System'.