Update convert-classic-resource.md
I recently worked on the Support Request 2308150040009508, where the customer requested me to provide a power shell script that helps them to identify the application insights resources based on the ingestion types. (Classic or workspace based)
For customers who have over 50+ resources, it is hard to locate each resource and identify which ones are classic and which ones are workspace based. The power shell script helps them to identify the same so they can prepare for a seamless migration.

Please make sure to add my name as a contributor on the public doc

Thank you in advance,