Update how-to-upgrade-previous-version.md
Its conflicting how in-place states the a Pro is it does not require another server but then also states as a Con that it does? I removed it as a Con as reading online other have said its not a requirement but leave it up to you guys to have the final say it is to be a pro or con. Am I mistaken and its also both?