(AzureCXP) fixes MicrosoftDocs/azure-docs#115311
We made chnages in line no: 55

From: 
Groups can be role-assignable or non-role-assignable. The group can be enabled in PIM for Groups or not enabled in PIM for Groups.
To:
Groups in Azure AD can be classified as either role-assignable or non-role-assignable. Additionally, any group can be enabled or not enabled for use with Azure AD Privileged Identity Management (PIM) for Groups