Update regex in how-to-rename-azure-ad.md
Various regex fixes

The pattern `[a-z|0-1]*` only matches letters and `|`, `0` and `1`. It should be changed to `[a-z0-9/-]*`. I am not sure about the intention for `name="[^$]*`, but it currently matches an unlimited number of non `$` characters. Finally slashes (`/`) does not need to be escaped in PowerShell. I also limited the number of repeats, to reduce execution time.