Update how-to-develop-emulator.md
If the "LimitToEndpoint = true" configuration is not added to the CosmosClientOptions and the user is testing locally, the cosmos client never returns a response and just continually spins.  It appears that it just keeps trying to connect, but doesn't timeout correctly or has no limit.  This seems to fix that issue.