Update howto-authentication-passwordless-security-key-on-premises.md
I have provided alternate syntax which avoids the NTLM authentication failure when the PowerShell  is executed on a machine that is not in the same domain as the targeted domain for the cmdlet. Using -domaincredential $somecreds format where $somecreds is provided as domain\admin and password combination fails. The use of -userprincipalname is more likely to be of use as a majority of the admins will not be using this without having to do modern authentication for MFA.