Update faq.yml
Update line 170, as the one of the customer wants to add Universal Print is a replacement for Hybrid Cloud Print. Universal Print addresses additional scenarios, and is much easier to set up and scalable. Hence added to the line 170 with the same. 

Got this info from here - https://learn.microsoft.com/en-us/universal-print/fundamentals/universal-print-faqs