Update what-is-cloud-sync.md
Line 82 was referred to Active directory domain services, on checking this wiki, it should be Azure Active Directory Domain Services, Due to Entra we have changed the terminology of Azure ADDS to Microsoft Entra Domain Services. 

Hence updated the same and created a PR, please review the same and approve it. 

https://supportability.visualstudio.com/AzureAD/_wiki/wikis/AzureAD/236881/Azure-AD-Cloud-Sync?anchor=comparison-between-azure-ad-connect-and-cloud-sync