Update connectivity-settings.md
As working with multiple customers about below user scenario:

Synapse workspace integrates with Git repo, and public network access is disabled.

Under this situation, we found user can still access to the workspace, it "bypass" the network setting due to studio directly calls Git api. And user can commit, create pull request, create branch and delete branch, but cannot publish changes to synapse live mode.

As discussed with Synapse Studio PM, currently this is by design and need to have this important notes in the public doc. So request this commit to merge the important notes.