Update analyze-cost-data-azure-cost-management-power-bi-template-app.md
If you leave the Scope "as is," it is only displaying a label. Therefore, the field would have no value and the refresh fails. Per your troubleshooting below, you need to enter the text. So, I believe step 9 can be better understood this way.

I believe the image will need to be updated as well.