Update azure-CA-details.md - update RSA 2017 CA to the 2019 one
There are two Microsoft RSA 2017 CA "not before 2017" and "not before 2019". The earlier one is archived (using /archived/ url) and is no longer issuing, the other one is active one and currently issuing. The first tab of documentation points at the archived cert. The second tab of chains points at the active cert.

Update the first documentation tab to point at the correct cert.