(AzureCXP) fixes MicrosoftDocs/azure-docs#117391
I have tested this in my local environment using the existing cmdlet in .yml to install the dotnet entity framework the github action is getting failed as shown below. 

As user suggested, when I added the --version parameter to install dotnet entity framework in step 7.2 able to run and deploy the code to app service.