Update prometheus-sidecar-remote-write-managed-identity-yaml.md
Ensured the self-managed Prometheus was correct created, and all pods are working well.
 
If I use this exactly yaml file "https://learn.microsoft.com/en-us/azure/azure-monitor/containers/prometheus-remote-write-managed-identity#deploy-side-car-and-configure-remote-write-on-the-prometheus-server", just copy all content and insert cluster information the pod is updated but the container “prom-remotewrite” is not created.
   
No error messages. Theoretically, the deployment was finished correctly.

Please follow the yaml sample with indentation corrections.