Create record-calls-locally-web.md
Local Recording feature allows a teams user to record a meeting and save the recording locally to their devices. This is giving the ACS developer the ability to show notification when such recording is started or stopped by a teams user. These are the how-to instruction to help the developers get started with using the feature.