Update PS custom role definition name
In the PowerShell cmdlet to create the custom role using the JSON definition, it previously used the name "Key Vault resource manager template deployment operator" as a value for the parameter "-RoleDefinitionName" which did not reflect what was provided in the JSON definition (e.g. "Key Vault Bicep deployment operator"). 

In my edit, I changed the value of the parameter to match what is written in the JSON definition. This will help create consistency for those following the tutorial using the copy/paste operations.