Update concept-aws-connector.md
change principle for principal, change signing validation for token signature validation, since Entra ID issues a token with a digital signature that AWS validates before exchanging it for the AWS set of credentials used by CSPM  to scan AWS inventory. Image with the architecture also has mispellings and should be changed, not on this commit.