echo back the given 'name' query parameter
As it is now, the doc says that the given 'name' query parameter is echoed back in the response. It is not. I remember in the past this sample code did echo it back, but for whatever reason it doesn't now - so the code is out of sync with the doc. I propose this code change to make it back in sync with the doc.