Update subnet creation in S2S VPN tutorial for Azrue VPN Gateway
Update the article "Tutorial - Connect an on-premises network and a
virtual network: S2S VPN: Azure portal - Azure VPN Gateway" to correct
the process of the gateway subnet creation.

The current tutorial suggests to create the gateway subnet as part of
the gateway creation. The suggested configuration is not shown in the
Azure Portal.

The update suggests to adjust the subnet creation to be an additional
step after the creation of the VNet.

Resolves #118572