Update immutable-storage-overview.md
If there is version level Immutable policy configured on the account level, the inventory report tends to fails as it is a conflicting feature. Additionally, if there an immutable policy configured on the destination containers of the report, then also the report shall tend to fail.

Adding both the scenarios on this documentation too and in the same request for review.