Fix description of ContinueOnError (PR MicrosoftDocs#263363)
This is a commit regarding the following review.

-------------------------------------------------------------------
kengaderdus 3 days ago
Can we use a NOTE just below the table instead? Something like:

[NOTE]
Currently, if you set ContinueOnError to false for a validation technical profile of type ClaimsTransformation, the technical profile doesn't honor the setting. To overcome this issue, use Preconditions instead.