Update network-requirements.md
Can you please see if this statement can be elaborated. Field engineers depending on their research sequence if they land up on this page it seems so that this note is making a contradictory statement to what is listed here.

https://learn.microsoft.com/en-us/azure/sentinel/best-practices-data#on-premises-windows-log-collection

Row - Servers do not connect to the internet Use the Log Analytics gateway	Configuring a proxy to your agent requires extra firewall rules to allow the Gateway to work.

Note : If outbound connectivity is restricted by your firewall or proxy server, make sure the URLs and Service Tags listed below are not blocked.