Fix the tutorial to delete an ARO cluster
The description of the command to delete the ARO cluster is not accurate:

> az aro delete --resource-group $RESOURCEGROUP --name $CLUSTER
>
> You'll then be prompted to confirm if you want to delete the cluster. After you confirm with y, it will take several minutes to delete the cluster. When the command finishes, the entire resource group and all resources inside it, including the cluster, will be deleted.

It is not true because that command only deletes the cluster. The virtual network and the resource group will not be touched. 

This commit modifies the tutorial by suggesting deleting the entire resource group when user wants to delete the cluster. Doing so, it is also aligned with the AKS tutorial.