Update deployment-script-template.md
Azure Storage account cannot be created in a private network. Instead we allow traffic from private network to storage account using service endpoint. This is different from services that are part of the vnet. So it would be great to rephrase "create a storage account in a private network" as "create a storage account". The section below talks about how to integrate the storage account with a VNET, so the meaning should still be clear.