Update configure-expressroute-private-peering.md
On below page 

https://learn.microsoft.com/en-us/azure/expressroute/configure-expressroute-private-peering#enable-private-peering

The statement about the shared key asks to Enter MD5 hash for the shared key.

"Shared key: Enter an MD5 hash for the shared key. The use of a shared key is optional."

If we calculate the MD5 hash for shared key "Microsoft" the MD5 hash value is - "140864078aeca1c7c35b4beb33c53c34"

This is 32 character long. while maximum permissible character to use for Shared key is 25 characters. So we cannot add the MD5 hash in shared key. Also from security point of view hash should be calculated by both end on the Password phrase.