Update modeling-simulation-workbench-overview.md
We are preview, and a form submission is required to request feature flag to use our service. Making that available for new customers onboarding to find in our overview doucmention.

Also we have a support feedback email available, and needed to publish our regions which are available as part of the preview.