Fix KeyError by correcting image_file_id index
Changed the indexing in the assignment of image_file_id to use the first item in the data list. Previously, using the second item (data['data'][1]) caused a KeyError because the correct image file ID is located in the first item (data['data'][0]). This update ensures the proper retrieval of the image_file_id, allowing the function to operate correctly.