Update custom-domain.md to remove flawed sample and replace with instructions to request a support ticket
We were told by customers that it is possible to bypass the sample in docs that talks about blocking access to the b2clogin domain. This change removes it and replaces it with instructions to open a support ticket if the customer wants it since B2C team has a backend flag to enable per-customer on demand.