On branch edburns-msft-em-4966-liberty-aro-quickstart-reza modified:   articles/openshift/howto-deploy-java-liberty-app.md
```
Under Instance details, select the region for the deployment – there is a step missing after this.
Use the same value for Confirm secret - there is a step missing after this.
```

modified:   articles/openshift/includes/aro-quota.md

```
This needs to say a little more about how to increase quota to run ARO.
```

Signed-off-by: Ed Burns <edburns@microsoft.com>