Update monitor-instances-health-check.md
We're trying to differentiate between when the problem is caused by customer's Web App's code and when the problem is caused by the instance (Web Worker) itself, as customers often assume the problem is the instance itself and not their code. This not only creates the false perception that the platform's instances are unstable, but also it causes unnecessary requests from customers to support to remove the "bad instance".