Audit diagnostic settings parameter info
Azure customers are finding the `Audit diagnostic setting for selected resource types` build in policy difficult to use holistically, as there isn't a well-defined way to populate the `listOfResourceTypes` parameter for all supported resourceTypes.

This change includes a bash script to help with this. Would be happy to do a PowerShell equivalent too if useful?