Update azure-monitor-agent-windows-client.md
Updated PowerShell onboarding script as it was not working as written. The $user variable line did not produce any results for $user variable which made any references to the $user variable downstream fail. Modified line to:

$user = Get-AzADUser -SignedIn

Added a few modifications, resolved spelling mistakes and added a couple of comments. Added #-UseDeviceAuthentication to the end of Connect-AzAccount cmdlet to accommodate for an error exhibited in testing.

Used PowerShell multiline comment for optional content, should user want to use optional content they can uncomment the block.