Add https://aka.ms/migratedisksfromLRStoZRS as one of the next steps
Hi @beavincent , I'm Penny Ko. I'm proposing to add the https://aka.ms/migratedisksfromLRStoZRS link as one of the next steps to this doc.

The reason is because for APRL, we are being asked to include only 1 resource link to align with Azure Advisor recommendations. I hope to make it easier for customers to find the how to convert doc. Thanks!