Update components.md
The following two facts haven't been mentioned anywhere in the documentation. They are among questions asked in Azure certification exam.

"VMs do not need a public IP address in order to be attached to backend pool of a public load balancer. Also, you can attach VMs to the backend pool of a load balancer even if they are in a stopped state. "