Clarification to disable/delete DDoS - manage-ddos-protection.md
Received feedback from CSS that there are a lot of customers reaching out for refunds because of DDoS costs when they are not using the DDoS protection on any virtual networks. They said that the documentation is not clear as to what disabling the DDoS protection means and how to fully delete the plan. 

To solve this, I added a NOTE to the Disable from a virtual network section, explaining what it means better.